+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+# [1.4.0] - 2021-02-27
+
+-   feat: commands for sending arbitrary input to matron/crone
+-   feat: script reload command
+-   feat: sleep command
+
 # [1.3.1] - 2021-02-14
 
 -   fix: correct cursor position when pasting mid-buffer